[feat/home_web_view]: slideUp, slideDown 애니메이션 추가
