+++ presentation/src/main/java/org/gdsc/presentation/utils/viewExtensions.kt
@@ -3,8 +3,6 @@ package org.gdsc.presentation.utils
 import android.text.Editable
 import android.text.TextWatcher
 import android.view.View
-import android.view.animation.AlphaAnimation
-import android.view.animation.Animation
 import android.widget.EditText
 
 fun EditText.addAfterTextChangedListener(block: (String) -> Unit) {
@@ -19,36 +17,6 @@ fun EditText.addAfterTextChangedListener(block: (String) -> Unit) {
         override fun onTextChanged(s: CharSequence?, start: Int, before: Int, count: Int) {
         }
     })
-}
-
-fun View.fadeIn() {
-    val self = this
-    val animation = AlphaAnimation(0f, 1f)
-    animation.duration = 200
-    animation.setAnimationListener(object : Animation.AnimationListener {
-        override fun onAnimationStart(animation: Animation) {
-            self.visibility = View.VISIBLE
-        }
-        override fun onAnimationRepeat(animation: Animation) {}
-        override fun onAnimationEnd(animation: Animation) {
-        }
-    })
-    self.startAnimation(animation)
-}
-
-fun View.fadeOut() {
-    val self = this
-    val animation = AlphaAnimation(1f, 0f)
-    animation.duration = 200
-    animation.setAnimationListener(object : Animation.AnimationListener {
-        override fun onAnimationStart(animation: Animation) {
-            self.visibility = View.GONE
-        }
-        override fun onAnimationRepeat(animation: Animation) {}
-        override fun onAnimationEnd(animation: Animation) {
-        }
-    })
-    self.startAnimation(animation)
 }
 
 fun View.getAbsolutePositionOnScreen(): Pair<Float, Float> {
